refactor: move common fns up
